refactor(hooks): move signup-related hooks to services directory
Reorganized signup-related hooks by moving them from the hooks directory
to the services directory for better code structure and maintainability.
