+++ src/services/signup/submit.ts
@@ -1,13 +1,52 @@
-import { signup } from "@/api/user/user-service";
-import { inputEmailCheck } from "@/utils/input-check/input-email-check";
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { useSignupContext } from "@/hooks/use-signup-context";
 import { inputNullCheck } from "@/utils/input-check/input-null-check";
+import { inputEmailCheck } from "@/utils/input-check/input-email-check";
+import { signup } from "@/api/user/user-service";
 import { storage } from "@/utils/storage/user-storage";
-
 
 interface SignupResult {
   success: boolean;
   errors: string;
 }
+
+export const useSignupSubmit = () => {
+  const navigate = useNavigate();
+  const signupAuth = useSignupAuth;
+  const [isSignupLoading, setIsSignupLoading] = useState(false);
+  const [textErrorMessage, setTextErrorMessage] = useState("");
+
+  const {
+    name,
+    email,
+    password,
+    confirmPassword,
+    checkTrigger,
+    setCheckTrigger
+  } = useSignupContext();
+
+  const handleSignup = async () => {
+    setCheckTrigger(!checkTrigger);
+    setIsSignupLoading(true);
+    
+    const signupResponse = await signupAuth(name, email, password, confirmPassword);
+    if (signupResponse.success === true) {
+      setTextErrorMessage("");
+      navigate("/roomlist");
+    } else {
+      setTextErrorMessage(signupResponse.errors);
+    }
+    
+    setIsSignupLoading(false);
+  };
+
+  return {
+    isSignupLoading,
+    textErrorMessage,
+    handleSignup,
+  };
+};
 
 export const useSignupAuth = async (
   inputName: string,
